fragmennts

+++ app/src/main/java/ru/netology/nmedia/repository/PostRepository.kt
@@ -5,8 +5,8 @@ import ru.netology.nmedia.dto.Post
 
 interface PostRepository {
     fun getAll(): LiveData<List<Post>>
-    fun like(id: Int)
-    fun repost(id: Int)
-    fun removeById(id: Int)
+    fun like(id: Long)
+    fun repost(id: Long)
+    fun removeById(id: Long)
     fun save(post: Post)
 }
